refactor(shared-blend): move shared time -and frequency format logic to wrapper
