--- src/scripts/components/stories/story/blocks/image-gallery/formats/frequency-blend/frequency-blend-image/frequecy-blend-image.module.css
@@ -1,12 +1,0 @@
-.frequencyImageBlend {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  object-fit: cover;
-
-  li {
-    list-style-type: none;
-  }
-}
