+++ src/scripts/components/stories/story/blocks/image-gallery/formats/shared/blend-image/blend-image.module.css
@@ -1,4 +1,4 @@
-.timeImageBlend {
+.blendImage {
   position: absolute;
   top: 0;
   left: 0;
