--- src/scripts/components/stories/story/blocks/image-gallery/formats/time-blend/time-blend.module.css
@@ -1,56 +1,0 @@
-.stickySectionWrapper {
-  position: relative;
-
-  .stickyScroller {
-    height: calc(100dvh - var(--header-height));
-    position: sticky;
-    top: 0;
-    overflow: hidden;
-    /* Keep content from spilling out */
-
-    .imageContainer {
-      position: relative;
-      /* New container for absolute positioning */
-      width: 100%;
-      height: 100%;
-
-      li {
-        list-style-type: none;
-      }
-    }
-
-    .altText {
-      height: 4.375rem;
-      width: 100%;
-      position: absolute;
-      bottom: 0;
-      left: 50%;
-      transform: translateX(-50%);
-      background-color: var(--Neutral-ESA-Background, #011e2b);
-      padding: 16px 24px;
-      color: var(--Neutral-ESA-White, #fff);
-      text-shadow:
-        0px 1px 2px rgba(0, 0, 0, 0.3),
-        0px 1px 3px rgba(0, 0, 0, 0.15);
-      font-family: NotesEsa, sans-serif;
-      font-size: 0.875rem;
-      font-style: normal;
-      line-height: 1.25rem;
-      letter-spacing: 0.175px;
-
-      >* {
-        text-overflow: ellipsis;
-        white-space: nowrap;
-        overflow: hidden;
-      }
-
-      h3 {
-        font-weight: 700;
-      }
-    }
-
-    p {
-      font-weight: 400;
-    }
-  }
-}
